requirements.txt
absl-py==0.13.0
aiohttp==3.7.4.post0
antlr4-python3-runtime==4.8
async-timeout==3.0.1
attrs==21.2.0
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.4
fsspec==2021.7.0
future==0.18.2
google-auth==1.35.0
google-auth-oauthlib==0.4.5
grpcio==1.39.0
idna==3.2
kornia==0.5.10
Markdown==3.3.4
multidict==5.1.0
numpy==1.21.2
oauthlib==3.1.1
omegaconf==2.1.1
opencv-python==4.5.3.56
packaging==21.0
Pillow==8.3.1
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyDeprecate==0.3.1
pyparsing==2.4.7
pytorch-lightning==1.4.4
PyTurboJPEG==1.5.2
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.16.0
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
torch==1.9.0+cu111
torchaudio==0.9.0
torchmetrics==0.5.0
torchvision==0.10.0+cu111
tqdm==4.62.2
typing-extensions==3.10.0.2
urllib3==1.26.6
Werkzeug==2.0.1
yarl==1.6.3
